+++ tests/test_mods_off.rs
@@ -8,7 +8,11 @@ fn test_mods_off() {
     let handle: LoggerHandle = Logger::try_with_env_or_str("info, test_mods_off::mymod1=off")
         .unwrap()
         .format(detailed_format)
-        .log_to_file(FileSpec::default().directory(self::test_utils::dir()))
+        .log_to_file(
+            FileSpec::default()
+                .suppress_timestamp()
+                .directory(self::test_utils::dir()),
+        )
         .start()
         .unwrap_or_else(|e| panic!("Logger initialization failed with {}", e));