+++ tests/test_reconfigure_methods.rs
@@ -7,7 +7,11 @@ use log::*;
 fn test_reconfigure_methods() {
     let mut logger = Logger::try_with_str("info")
         .unwrap()
-        .log_to_file(FileSpec::default().directory(self::test_utils::dir()))
+        .log_to_file(
+            FileSpec::default()
+                .suppress_timestamp()
+                .directory(self::test_utils::dir()),
+        )
         .start()
         .unwrap_or_else(|e| panic!("Logger initialization failed with {}", e));