+++ tests/test_utils.rs
@@ -1,5 +1,6 @@
 #![allow(dead_code)]
 
+#[cfg(feature = "use_chrono_for_offset")]
 use chrono::{Local, Offset};
 use std::path::PathBuf;
 use time::{format_description::FormatItem, macros::format_description, OffsetDateTime, UtcOffset};